Ajustes genericos a listado de animes

+++ lib/models/media.dart
@@ -2,7 +2,7 @@ class Media {
 
   final int id;
   final String title;
-  final double score;
+  final int score;
   final String comment;
   final String category;
   final String status;
@@ -12,7 +12,7 @@ class Media {
   Media({
     this.id = 0,
     this.title = '',
-    this.score = 0.0,
+    this.score = 0,
     this.comment = '',
     this.category = '',
     this.status = '',
